Fix breakpoints & refactor TradingViewPopper
