+++ src/@demex-info/views/HeroSection/HeroSection.tsx
@@ -1,9 +1,9 @@
 import { BackgroundAnimation, CoinIcon, SvgIcon } from "@demex-info/components";
 import { getDemexLink, Paths } from "@demex-info/constants";
 import { eskimi, StyleUtils } from "@demex-info/utils";
-import { Box, Button, Container, Hidden, Typography, makeStyles, useMediaQuery, useTheme } from "@material-ui/core";
+import { Box, Button, Container, Hidden, makeStyles, useMediaQuery, useTheme } from "@material-ui/core";
 import clsx from "clsx";
-import React, { Suspense, useEffect, useRef } from "react";
+import React, { Suspense, useEffect } from "react";
 import { useInView } from "react-intersection-observer";
 import TextLoop from "react-text-loop";
 import { DesktopMobile, Mobile } from "./assets";
@@ -13,7 +13,6 @@ const HeroSection: React.FC = () => {
 	const classes = useStyles();
 	const theme = useTheme();
 	const [ready, setReady] = React.useState<boolean>(false);
-	const anchorEl = useRef(null);
 
 	const [titleRef, titleView] = useInView({
 		threshold: 0.2,
@@ -108,11 +107,8 @@ const HeroSection: React.FC = () => {
 					<Hidden smDown>
 						<Box className={classes.graphicsWrapper}>
 							<SvgIcon className={classes.svgIcon} component={DesktopMobile} />
-							<Box display="flex">
-								<div ref={anchorEl} className={classes.tradingViewWrapper}>
-									<Typography className={classes.tradingViewText}>Charts by </Typography>
-									<TradingViewPopper anchorEl={anchorEl} />
-								</div>
+							<Box className={classes.tradingViewWrapper}>
+								<TradingViewPopper />
 							</Box>
 						</Box>
 					</Hidden>
@@ -120,10 +116,7 @@ const HeroSection: React.FC = () => {
 					<Hidden mdUp>
 						<Box className={classes.mobileGraphicsWrapper}>
 							<SvgIcon component={Mobile} />
-							<div ref={anchorEl} className={classes.tradingViewWrapper}>
-								<Typography className={classes.tradingViewText}>Charts by </Typography>
-								<TradingViewPopper anchorEl={anchorEl} />
-							</div>
+							<TradingViewPopper />
 						</Box>
 					</Hidden>
 
@@ -404,25 +397,27 @@ const useStyles = makeStyles((theme) => ({
 		flexDirection: "column",
 		alignItems: "center",
 	},
-	tradingViewText: {
-		...theme.typography.body3,
-		color: theme.palette.text.disabled,
-		fontWeight: "bold",
-	},
 	tradingViewWrapper: {
-		display: "flex",
-		justifyContent: "end",
 		position: "absolute",
 		bottom: "5%",
 		right: "20%",
+		"@media only screen and (min-width: 1720px) and (max-width: 1920px) ": {
+			right: "25%!important",
+		},
+		"@media only screen and (min-width: 1630px) and (max-width: 1720px) ": {
+			right: "22%!important",
+		},
+		[theme.breakpoints.down("xl")]: {
+			// right:"25%",
+		},
 		[theme.breakpoints.down("lg")]: {
 			right: "15%",
 		},
 		[theme.breakpoints.down("md")]: {
-			bottom: "10%",
+			bottom: "5%",
 		},
-		[theme.breakpoints.down("sm")]: {
-			position: "static",
+		"@media (max-width: 1000px)": {
+			bottom: "15%",
 		},
 	},
 	svgIcon: {
